Extract catalyst logic to another file
